Disabled responsive images in the backend for the time being.
Explanation: there's a CSS technique for making images responsive by applying
max-width: 100%; and height: auto; to them. This messes with the bundled Xinha
WYSIWYG editor. The buttons in Xinha's toobars can't have max-width: auto;, it
makes them disappear.

So far, I haven't found a way to overwrite this for the toolbar images only due
to selector specifity. Either need to find a way or find a way to make other
images (media, actually) in the backend have a class which assigns the responsive
media CSS to them.

(Or drop Xinha for good, but that doesn't seem to be an option.)