Fix the error_reporting somewhat; until now, prevent E_STRICT errors were not correctly surpressed, but we need to do this, because otherwise older plugins will prevent Serendipity from loading.
I actually wanted to at least "break" stuff and report errors, when $serendipity['production'] is set to "debug".

However, this does NOT work for me (PHP 5.2); even when set to "debug", E_STRICT errors do not seem to show up.
I'm quite at a loss here, so anyone reading this, go ahead and inspect.

At least it makes this version work when "older" plugins exists. It would be great if the code could be made to show the debug errors in debug mode.

I shuffled around the error_reporting() position because when someone sets the variable in serendipity_config_local.inc.php it would be too early to affect code flow; now it should evaluate the variable properly.