Dashboard: open further links in a modal popup.
(Or just show/hide them for use-popups.)

For some weird reason, this seems to fire only on the second click
on the toggler, not on the first. Need input from @onli on this.

Also, this needs a better icon. I did not want to go through a new
iconfont build just for this one icon now.