Only replace T in value if datetime-local is not supported.
According to my tests, this is what makes datetime-local fail in
browsers which already support it. Replacing the T in the value is
required to set the date/time in browsers which don't support it,
but it somehow throws off the (re)setting in browsers which do.

So using Modernizr's feature test to check if the replacement is
required should solve the issue. However, this seems strange, we
should keep an eye on it.