Make category filtering a reusable function for live filtering
Can now be used anywhere where a "real-time" filter is desirable.
As a usage example, it's implemented for the list of installable
plugins. Should work anywhere, though – needs an input element,
a list of target elements and an element within those target
elements to contain the search term.

Also made the function for the reset button for this reusable.

References #154