Init basic Smarty when requesting CSS (Fixes #204)
In order to be able to call `serendipity_plugin_api_pre_event_hook` in a template's config.ini.php from serendipity.css.php, Smarty must be initialized first.
In case of a 'css' event, within the event hook $eventData contains the template's CSS which can now be manipulated and/or extended.