extends 9cbb2b0
for file renaming type='file' with a pdf also. Thumb check and the silencer were both needed, else for the last the database entry was lost.

The $renameData['thumb'] could very well be $file['thumbnail_name'] in case of not being empty.

It might be that this has dependencies with the hotlink hook or thumbSuffix renamings. Garvin, please check.

References #220