Sync heights of info overlays
This has been an optical issue for both media files and themes,
which is why my original idea to have large preview files would
not have solved it. However, a CSS-based solution would not have
been solid, either, so I went with a few lines of jQuery to sync
the height with the outer container.

References #230