A better solution for media file/template info overlays
Since the JS-based solution does not work well for cases where the
overlay is larger than the (very small) image, we'll have to go
with a CSS-based solution. This sets the height of the overlay
to at least 7em, which should cover the core info of the overlay
(which is usually 4 lines of text). Also, we'll have to go with
overflow: scroll; in case the info is longer.

Not a good solution in a reponsive context, but it should work. :(

References #230