fix nasty bug with high PHP versions
@garvinhicking
May this be the "real" cause of http://board.s9y.org/viewtopic.php?f=10&t=20639 ff, not loading the editor.js file? He is using PHP 5.6.17.
What I do not get on this: Why does this only happen with the Standard Editor on his site? And why does this not happen testing both groups to me locally with nearly the same PHP version?