[spamblock] Remove dead code (Akismet spam toggle)
This code was used to tell Akismet that a comment
is spam or ham. It can't be triggered as there
are no UI buttons, and it shouldn't be used as
it uses GET params. I don't think it will be
missed.

The relevant issue has been open for about two
and a half years without a new implementation,
so let's drop it. We can always re-implement it
later ...

Closes #405.

Signed-off-by: Thomas Hochstein <thh@inter.net>