Revert "Fix pagination for "stable archives" in themes."
This reverts commit 4e662bf17c420cc997c39c9501cfe0df509d9a09
in parts.

With this - now reverted - commit, the prev/next
arrows in pagination were swapped because they
were the wrong way around. Unfortunately, later on,
the root cause was fixes by "unswapping" the prev/next
links in core, so now the pagination arrows were
... eh ... wrongly swapped. So this commit has to
be reverted.

Backported from master branch.

Signed-off-by: Thomas Hochstein <thh@inter.net>