Change meaning of PLUGIN_DOCUMENTATION.
The "website" property fo the plugin bag used
to contain a link to external documentation,
but there is none left. So in future it may
point to a plugin website (or the website
of the plugin's author, if there is no
special page about the plugin).

Change language and translations accordingly.

Signed-off-by: Thomas Hochstein <thh@inter.net>