XML/RPC: Fix spl_object_hash parser conversion (#875)
* Replace (int) parser conversion with spl_object_hash
The (int) conversion does not work anymore since PHP 8 changed the xml parser from being a resource to being an object. Hardcoding it to a number, the initial fix, destroys the logic of having multiple parsers available and was not done everywhere. The object hash approach should be a valid replacement.

* Add test for XML_RPC_Message

* Add test for XML_RPC_Response, when valid

* Add test for invalid XML_Response, simplify code

* Fix $message access

* Try test fix for PHP 7.4

* Try test for PHP 7.4, LANG-CHARSET missing

* PHP 7.4 bc fix for XML_PARSER hash

* Try PHP 7.4 'ha' fix

* Try fix for undefined 'ha' index

* Use ?? check for legibility

Co-authored-by: Garvin Hicking <blog@garv.in>

---------

Co-authored-by: Garvin Hicking <blog@garv.in>