Fix tooltip and language listing in installer (#918)
* Fix tooltips in installer
The JS file in the installer is coming from default, which was not updated for the new jquery version. size needed to be replaced with length and a load event rewritten.

* Fix language listing in installer
By setting UTF-8 as default charset during the installer run. The language list is encoded in UTF-8, but the charset set in the header was still ISO-8591

* document changes