Show proper error message when uploaded file is too big for PHP (#926)
* Show error messages when images are too big for the PHP settings
Detects the situation when 'post_max_size' was the problem and prints a separate error then. Alternatively upload_max_filesize can be the problem. Here however the existing flow already presented a proper error message, this was just broken prior by setting the 500 error header, which made the ajax uploader react properly to errors. Fixing that bug for the PHP uploader restored the error message, so that the upload_max_filesize scenario is also covered.

* Move upload error message to MEDIA_UPLOAD_POST_MAX_SIZEERROR

* document changes