Brute force protection for serendipity login (#927)
* Store cache ttl with cache item
This enables us to use the cache on other scenarios than just the entries cache (with its fixed duration). Also disables the redundant check for the cache option, since the code in functions_entries already checks this.

* Brute force protection for user accounts
Allow 5 login tries a minute, block for a minute

* Rate limit logins by IP
Prevent a potential attack where futile login attempts make it impossible for the real user to log in.

* document changes

* minor: adjust function comment

* Fix: Browsing the backend was counted as failed log-in