clean blog: Fix wide navbar floating far below the top (#939)
* clean blog: Fix wide navbar floating far below the top
When the navigation had many or long links, the right part gets placed below the left part of the header. In that case the translate3d approach did not work anymore, instead of the whole header appearing at the very top of the viewport it appeared with a big gap further below. Just setting the top property to 0 instead fixed this behaviour in my tests.

* document changes